✨[feat]: NetworkAble 프로토콜 정의 및 MockTestData 객체 구현
