+++ OpenMarket/OpenMarket/Model/Network.swift
@@ -1,0 +1,16 @@
+//
+//  Network.swift
+//  OpenMarket
+//
+//  Created by 우롱차, Donnie on 2022/05/10.
+//
+
+import Foundation
+
+protocol NetworkAble {
+    func inquireProductList() -> Data?
+}
+
+class Network {
+    
+}
