+++ OpenMarket/OpenMarketJSONTests/OpenMarketJSONTests.swift
@@ -2,31 +2,31 @@
 //  OpenMarketJSONTests.swift
 //  OpenMarketJSONTests
 //
-//  Created by SeoDongyeon on 2022/05/10.
+//  Created by 우롱차, Donnie on 2022/05/10.
 //
 
 import XCTest
 @testable import OpenMarket
 
-class OpenMarketJSONTests: XCTestCase {
+final class OpenMarketJSONTests: XCTestCase {
     private var jsonParser: JsonParser!
+    private var mockTestData: NetworkAble!
     
     override func setUpWithError() throws {
         try super.setUpWithError()
         jsonParser = JsonParser()
+        mockTestData = MockTestData()
     }
     
     override func tearDownWithError() throws {
         try super.tearDownWithError()
         jsonParser = nil
+        mockTestData = nil
     }
     
     func test_products_JSON_테스트파일을_decoding해서_결과는_NotNil() {
         // given
-        let testFileName = "PageInformationTest"
-        let extensionType = "json"
-        
-        guard let data = load(fileName: testFileName, extensionType: extensionType) else {
+        guard let data = mockTestData.inquireProductList() else {
             XCTFail()
             return
         }
@@ -37,17 +37,6 @@ class OpenMarketJSONTests: XCTestCase {
         // then
         XCTAssertNotNil(pageInformation)
         XCTAssertNotNil(pageInformation?.pages)
-    }
-    
-    private func load(fileName: String, extensionType: String) -> Data? {
-        let testBundle = Bundle(for: type(of: self))
-        guard let fileLocation = testBundle.url(forResource: fileName, withExtension: extensionType) else { return nil }
-        do {
-            let data = try Data(contentsOf: fileLocation)
-            return data
-        } catch {
-            return nil
-        }
     }
 }
